Fixing modals

+++ src/views/managers/group.svelte
@@ -11,7 +11,6 @@
     NameDescModal,
     WhereIsItUsedModal,
     Spinner,
-    RemoveManagerModal,
   } from '../../components';
   import {
     GeneralService,
@@ -83,14 +82,24 @@
     );
   }
   async function remove() {
-    await GeneralService.errorWrapper(
-      async () => {
-        await EntityManagerService.delete('group', group._id);
-      },
-      async () => {
-        NotificationService.success('Group was successfully deleted.');
-      }
-    );
+    if (
+      await ConfirmService.confirm(
+        `Delete ${group.label} Group`,
+        `Are you sure you want to delete <strong>${group.label}</strong> group?
+        If deleted, the group will be removed from all templates, widgets
+        and entries that are using it.`,
+        group.name
+      )
+    ) {
+      await GeneralService.errorWrapper(
+        async () => {
+          await EntityManagerService.delete('group', group._id);
+        },
+        async () => {
+          NotificationService.success('Group was successfully deleted.');
+        }
+      );
+    }
   }
   async function addProp(prop: Prop) {
     await GeneralService.errorWrapper(
@@ -256,7 +265,7 @@
               updateProp(event.detail);
             }}
             on:deleteEntity={() => {
-              StoreService.update('RemoveManagerModal', true);
+              remove();
             }}
             on:deleteProp={(event) => {
               removeProp(event.detail);
@@ -308,13 +317,4 @@
       whereIsItUsedItems = [];
     }} />
   <Spinner show={showSpinner} />
-  {#if group}
-    <RemoveManagerModal
-      title={`Delete "${group.label}" Group`}
-      warningMessage={`Are you sure you want to delete "${group.label}" group?
-  If deleted, the group will be removed from all templates, widgets and entries that are using it.`}
-      inputLabel="Confirm group name"
-      managerName={group.label}
-      on:done={remove} />
-  {/if}
 </Layout>
